+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adds beta `retrieveCarrierMetadata` function
 - Adds missing `api_keys` property to `User`
 - Improves Error Deserialization to dynamically handle edge cases that have a bad format
+- Adds a helper function to retrieve carrier metadata
 
 ## v6.1.0 (2023-04-04)